+++ electron/main/core/core.ts
@@ -4,14 +4,11 @@ import { syncManager } from './sync'
 
 class TrackingEngine {
     async start() {
-        console.log('[TrackingEngine] Initializing System Agent...')
         await sessionManager.start()
         syncManager.start()
-        console.log('[TrackingEngine] Running.')
     }
 
     async stop() {
-        console.log('[TrackingEngine] Stopping...')
         await sessionManager.stop()
         syncManager.stop()
     }